Course history view for the user
