+++ frontend/templates/frontend/history/history_compare_header.html
