+++ frontend/templates/frontend/history/history_compare_view.html
@@ -11,7 +11,6 @@
 
     <!-- Show fields differences -->
         {% for field_diff in compare_data %}
-            {{ compare_data }}
             <h3>
                 {% firstof field_diff.field.verbose_name field_diff.field.related_name %}
                 {% if field_diff.is_related and not field_diff.follow %}
