+++ frontend/urls.py
@@ -47,6 +47,9 @@ urlpatterns = [
             path('edit/',
                  views.course.EditCourseView.as_view(),
                  name='course-edit'),
+            path('history/',
+                 views.history.CourseHistoryCompareView.as_view(),
+                 name='course-history'),
             path('delete/',
                  views.CourseDeleteView.as_view(),
                  name='course-delete'),
@@ -91,19 +94,19 @@ urlpatterns = [
                  views.content.ContentReadingModeView.as_view(),
                  name='content-reading-mode'),
             path('<pk>/textfield-history/',
-                 views.content.TextfieldHistoryCompareView.as_view(),
+                 views.history.TextfieldHistoryCompareView.as_view(),
                  name='textfield-history'),
             path('<pk>/ytvideo-history/',
-                 views.content.YTVideoHistoryCompareView.as_view(),
+                 views.history.YTVideoHistoryCompareView.as_view(),
                  name='ytvideo-history'),
             path('<pk>/image-history/',
-                 views.content.ImageHistoryCompareView.as_view(),
+                 views.history.ImageHistoryCompareView.as_view(),
                  name='image-history'),
             path('<pk>/pdf-history/',
-                 views.content.PdfHistoryCompareView.as_view(),
+                 views.history.PdfHistoryCompareView.as_view(),
                  name='pdf-history'),
             path('<pk>/latex-history/',
-                 views.content.LatexHistoryCompareView.as_view(),
+                 views.history.LatexHistoryCompareView.as_view(),
                  name='latex-history'),
 
         ])),
